Add files via upload
We are using UDP (User Datagram Protocol) over client-server
paradigm to achieve the functionalities of this portal. Every
individual seller and buyer will have a .txt file stored in the Server
(Admin). There is a sign up form, in which both Sellers and
Buyers will have to specify whether they are seller or buyer, when
creating their profile. There will also be individual file for each
category of products such as electronics, furniture, books, sports
etc.
If a client wants to buy a phone, say MI3, he will have to send a
request to server which is the ecommerce site. The server will
provide all the categories of products. The client will pick
Electronics Categories. The Electronics Category file contains all
the products as well as their sellers. The client can choose any
product from any seller. He can view all the details about the
product and seller. When the client buys the product, that
product’s quantity is decremented by 1 in both seller file and
category file to